Updated menus and added an options page.
