+++ Assets/Audio/Music.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 795426a19f01ec24c9a8ed79d359309d
+folderAsset: yes
+timeCreated: 1512860965
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
