+++ Assets/Audio/Music/04 Walk This Way.mp3.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: e90235f98e5b1db4eb527499dd849f65
+timeCreated: 1512865092
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
