+++ Assets/Audio/Music/A Warm Hypnosis.mp3.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 017bf3159c453f345877a3bde4c5b08e
+timeCreated: 1512865244
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
