+++ Assets/MenuActions.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1f9a62983f6ea4349b277eda40e05ebb
+timeCreated: 1512944355
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
