+++ savegame.data
@@ -2,7 +2,7 @@ Saving the Game
 PlayerPosX=328
 PlayerPosY=140.1005
 PlayerPosZ=174.2
-PlayerRotateX=-2.983481E-16
-PlayerRotateY=7.379185E-10
+PlayerRotateX=-8.730066E-16
+PlayerRotateY=2.166729E-09
 PlayerRotateZ=0
 load10coins
